+++ manta-trusted-setup/src/groth16/mpc.rs
@@ -18,22 +18,24 @@
 
 use crate::{
     groth16::kzg::{self, Accumulator},
-    ratio::{HashToGroup, RatioProof},
     util::{batch_into_projective, batch_mul_fixed_scalar, merge_pairs_affine},
 };
 use alloc::{vec, vec::Vec};
 use ark_groth16::{ProvingKey, VerifyingKey};
 use ark_poly::{EvaluationDomain, Radix2EvaluationDomain};
-use core::clone::Clone;
 use manta_crypto::{
     arkworks::{
-        ec::{AffineCurve, ProjectiveCurve},
+        ec::{AffineCurve, PairingEngine, ProjectiveCurve},
         ff::{Field, PrimeField, UniformRand, Zero},
         pairing::{Pairing, PairingEngineExt},
+        ratio::{HashToGroup, RatioProof},
         relations::r1cs::{ConstraintSynthesizer, ConstraintSystem, SynthesisError},
     },
     rand::{CryptoRng, RngCore},
 };
+
+#[cfg(feature = "serde")]
+use manta_util::serde::{Deserialize, Serialize};
 
 /// Proving Key Hasher
 pub trait ProvingKeyHasher<P>
@@ -52,6 +54,62 @@ pub type State<P> = ProvingKey<<P as Pairing>::Pairing>;
 
 /// MPC Proof
 pub type Proof<P> = RatioProof<P>;
+
+/// MPC State Size
+#[cfg_attr(
+    feature = "serde",
+    derive(Deserialize, Serialize),
+    serde(crate = "manta_util::serde", deny_unknown_fields)
+)]
+#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
+pub struct StateSize {
+    /// Size of `gamma_abc_g1` in the [`VerifyingKey`]
+    pub gamma_abc_g1: usize,
+
+    /// Size of the `a_query` in the [`ProvingKey`]
+    ///
+    /// # Note
+    ///
+    /// This is also the size of the `b_g1_query` and the `b_g2_query`.
+    pub a_query: usize,
+
+    /// Size of the `h_query` in the [`ProvingKey`]
+    pub h_query: usize,
+
+    /// Size of the `l_query` in the [`ProvingKey`]
+    pub l_query: usize,
+}
+
+impl StateSize {
+    /// Builds a [`StateSize`] from `proving_key` by measuring the lengths of the vectors in
+    /// `proving_key`.
+    #[inline]
+    pub fn from_proving_key<E>(proving_key: &ProvingKey<E>) -> Self
+    where
+        E: PairingEngine,
+    {
+        Self {
+            gamma_abc_g1: proving_key.vk.gamma_abc_g1.len(),
+            a_query: proving_key.a_query.len(),
+            h_query: proving_key.h_query.len(),
+            l_query: proving_key.l_query.len(),
+        }
+    }
+
+    /// Returns `true` if the lengths of the vectors in `proving_key` match `self`.
+    #[inline]
+    pub fn matches<E>(&self, proving_key: &ProvingKey<E>) -> bool
+    where
+        E: PairingEngine,
+    {
+        self.gamma_abc_g1 == proving_key.vk.gamma_abc_g1.len()
+            && self.a_query == proving_key.a_query.len()
+            && self.a_query == proving_key.b_g1_query.len()
+            && self.a_query == proving_key.b_g2_query.len()
+            && self.h_query == proving_key.h_query.len()
+            && self.l_query == proving_key.l_query.len()
+    }
+}
 
 /// MPC Error
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
@@ -88,8 +146,8 @@ pub enum Error {
 /// doing `eval_poly` in parallel on all four kinds of commitments that phase2 params require
 /// (`a_g1`, `b_g1`, `b_g2`, `extra`) where `extra` is the cross term that we (may wish to) compute
 /// separately for public/private inputs.
+#[allow(clippy::too_many_arguments)] // FIXME: Simplify this in the future.
 #[inline]
-#[allow(clippy::too_many_arguments)]
 pub fn specialize_to_phase_2<G1, G2>(
     tau_basis_g1: &[G1],
     tau_basis_g2: &[G2],
@@ -133,6 +191,24 @@ pub fn specialize_to_phase_2<G1, G2>(
                 }
             },
         );
+}
+
+/// Adds dummy `z_i * 0 = 0` constraint for each public input `z_i`.  This ensures non-malleability
+/// of Groth16 proofs even if some public inputs are otherwise unconstrained.
+#[inline]
+pub fn add_dummy_constraints<G>(
+    a: &mut [G],
+    ext: &mut [G],
+    tau_lagrange: &[G],
+    beta_tau_lagrange: &[G],
+    num_constraints: usize,
+    num_public_inputs: usize,
+) where
+    G: Clone,
+{
+    let total = num_public_inputs + num_constraints;
+    a[0..num_public_inputs].clone_from_slice(&tau_lagrange[num_constraints..total]);
+    ext[0..num_public_inputs].clone_from_slice(&beta_tau_lagrange[num_constraints..total]);
 }
 
 /// Checks that the parameters which are not changed by contributions are the same.
@@ -188,14 +264,9 @@ where
     constraints.finalize();
     let num_constraints = constraints.num_constraints();
     let num_instance_variables = constraints.num_instance_variables();
-    let domain = match Radix2EvaluationDomain::new(num_constraints + num_instance_variables) {
-        Some(domain) => domain,
-        None => return Err(Error::TooManyConstraints),
-    };
-    let constraint_matrices = match constraints.to_matrices() {
-        Some(matrices) => matrices,
-        None => return Err(Error::MissingCSMatrices),
-    };
+    let domain = Radix2EvaluationDomain::new(num_constraints + num_instance_variables)
+        .ok_or(Error::TooManyConstraints)?;
+    let constraint_matrices = constraints.to_matrices().ok_or(Error::MissingCSMatrices)?;
     let beta_g1 = powers.beta_tau_powers_g1[0];
     let degree = domain.size as usize;
     let mut h_query = Vec::with_capacity(degree - 1);
@@ -214,14 +285,14 @@ where
     let mut b_g1 = vec![C::G1::zero().into_projective(); num_witnesses];
     let mut b_g2 = vec![C::G2::zero().into_projective(); num_witnesses];
     let mut ext = vec![C::G1::zero().into_projective(); num_witnesses];
-    {
-        let start = 0;
-        let end = num_instance_variables;
-        a_g1[start..end]
-            .copy_from_slice(&tau_lagrange_g1[(start + num_constraints)..(end + num_constraints)]);
-        ext[start..end]
-            .copy_from_slice(&beta_lagrange_g1[(start + num_constraints)..(end + num_constraints)]);
-    }
+    add_dummy_constraints(
+        &mut a_g1,
+        &mut ext,
+        &tau_lagrange_g1,
+        &beta_lagrange_g1,
+        num_constraints,
+        num_instance_variables,
+    );
     specialize_to_phase_2(
         &tau_lagrange_g1,
         &tau_lagrange_g2,
@@ -241,15 +312,14 @@ where
     let ext = ProjectiveCurve::batch_normalization_into_affine(&ext);
     let public_cross_terms = Vec::from(&ext[..constraint_matrices.num_instance_variables]);
     let private_cross_terms = Vec::from(&ext[constraint_matrices.num_instance_variables..]);
-    let vk = VerifyingKey {
-        alpha_g1: powers.alpha_tau_powers_g1[0],
-        beta_g2: powers.beta_g2,
-        gamma_g2: C::g2_prime_subgroup_generator(),
-        delta_g2: C::g2_prime_subgroup_generator(),
-        gamma_abc_g1: public_cross_terms,
-    };
     Ok(ProvingKey {
-        vk,
+        vk: VerifyingKey {
+            alpha_g1: powers.alpha_tau_powers_g1[0],
+            beta_g2: powers.beta_g2,
+            gamma_g2: C::g2_prime_subgroup_generator(),
+            delta_g2: C::g2_prime_subgroup_generator(),
+            gamma_abc_g1: public_cross_terms,
+        },
         beta_g1,
         delta_g1: C::g1_prime_subgroup_generator(),
         a_query,
@@ -291,10 +361,7 @@ where
     R: CryptoRng + RngCore + ?Sized,
 {
     let delta = C::Scalar::rand(rng);
-    let delta_inverse = match delta.inverse() {
-        Some(delta_inverse) => delta_inverse,
-        _ => return None,
-    };
+    let delta_inverse = delta.inverse()?;
     batch_mul_fixed_scalar(&mut state.l_query, delta_inverse);
     batch_mul_fixed_scalar(&mut state.h_query, delta_inverse);
     state.delta_g1 = state.delta_g1.mul(delta).into_affine();